+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *Minor*
 
 * Added a Scapegoat role. ([#40](https://github.com/hjwylde/werewolf/issues/40))
+* Added a `status` command. ([#18](https://github.com/hjwylde/werewolf/issues/18))
 
 #### v0.3.2.0